requirements.txt
gunicorn==21.2.0
Django>=4.0,<5.0
djangorestframework>=3.14
django-cors-headers>=3.9
django-debug-toolbar>=3.2
requests>=2.25
